requirements.txt
Django==4.1.1
Pillow==9.3.0
django-filter ==21.1